ch02/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers~=3.10.0